Moved calculate_relocated_name from public_id to utils
